lots of style changes
