+++ hologram_dashboard/fetcher/hologram_data_source_availability_fetcher.cc
@@ -76,9 +76,4 @@ void HologramDataSourceAvailabilityFetcher::UpdateHistory(
     // TODO(alexanderlin): Implement.    
 }
 
-static std::unique_ptr<ProtoValue> ProtoValue::Create(std::string& text_proto, 
-    google::protobuf::Descriptor* descriptor) {
-    return nullptr;
-}
-
 } // namespace wireless_android_play_analytics
